+++ sphinx/application/network/concepts/queries/concept-network-query-contact/src/main/java/chat/sphinx/concept_network_query_contact/model/GetTribeMembersResponse.kt
@@ -1,0 +1,8 @@
+package chat.sphinx.concept_network_query_contact.model
+
+import com.squareup.moshi.JsonClass
+
+@JsonClass(generateAdapter = true)
+data class GetTribeMembersResponse(
+    val contacts: List<ContactDto>
+)