+++ sphinx/screens-detail/chat-detail/tribe-detail/src/main/java/chat/sphinx/tribe_detail/navigation/TribeDetailNavigator.kt
@@ -19,4 +19,6 @@ abstract class TribeDetailNavigator(
     )
 
     abstract suspend fun toCreateTribeScreen(chatId: ChatId)
+
+    abstract suspend fun toTribeMemberList(chatId: ChatId)
 }